% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depower-package.r
\docType{package}
\name{depower-package}
\alias{depower}
\alias{depower-package}
\title{depower: Power Analysis for Differential Expression Studies}
\description{
Provides a convenient framework to simulate, test, power, and visualize data for differential expression studies with lognormal or negative binomial outcomes. Supported designs are two-sample comparisons of independent or dependent outcomes. Power may be summarized in the context of controlling the per-family error rate or family-wise error rate. Negative binomial methods are described in Yu, Fernandez, and Brock (2017) \doi{10.1186/s12859-017-1648-2} and Yu, Fernandez, and Brock (2020) \doi{10.1186/s12859-020-3541-7}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brettklamer.com/work/depower/}
  \item \url{https://bitbucket.org/bklamer/depower/}
}

}
\author{
\strong{Maintainer}: Brett Klamer \email{code@brettklamer.com} (\href{https://orcid.org/0000-0002-8334-4831}{ORCID})

Authors:
\itemize{
  \item Lianbo Yu \email{lianbo.yu@osumc.edu} (\href{https://orcid.org/0000-0002-2025-2585}{ORCID})
}

}
\keyword{internal}
