% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_bnb.r
\name{lrt_bnb}
\alias{lrt_bnb}
\title{Likelihood ratio test for BNB ratio of means}
\usage{
lrt_bnb(data, ratio_null = 1, distribution = asymptotic(), ...)
}
\arguments{
\item{data}{(list)\cr
A list whose first element is the vector of negative binomial values
from sample 1 and the second element is the vector of negative
binomial values from sample 2.
Each vector must be sorted by the subject/item index and must be the
same sample size. \link[base]{NA}s are silently excluded. The default
output from \code{\link[=sim_bnb]{sim_bnb()}}.}

\item{ratio_null}{(Scalar numeric: \code{1}; \verb{(0, Inf)})\cr
The ratio of means assumed under the null hypothesis (sample 2 / sample 1).
Typically, \code{ratio_null = 1} (no difference). See 'Details' for
additional information.}

\item{distribution}{(function: \code{\link[=asymptotic]{asymptotic()}} or
\code{\link[=simulated]{simulated()}})\cr
The method used to define the distribution of the \eqn{\chi^2}
likelihood ratio test statistic under the null hypothesis. See
'Details' and \code{\link[=asymptotic]{asymptotic()}} or \code{\link[=simulated]{simulated()}} for
additional information.}

\item{...}{Optional arguments passed to the MLE function \code{\link[=mle_bnb]{mle_bnb()}}.}
}
\value{
A list with the following elements:
\tabular{llll}{
Slot \tab Subslot \tab Name \tab Description \cr
1 \tab \tab \code{chisq} \tab \eqn{\chi^2} test statistic for the ratio of means. \cr
2 \tab \tab \code{df}    \tab Degrees of freedom. \cr
3 \tab \tab \code{p}     \tab p-value. \cr
4 \tab \tab \code{ratio} \tab Estimated ratio of means (sample 2 / sample 1). \cr

5 \tab   \tab \code{alternative} \tab Point estimates under the alternative hypothesis. \cr
5 \tab 1 \tab \code{mean1} \tab Estimated mean of sample 1. \cr
5 \tab 2 \tab \code{mean2} \tab Estimated mean of sample 2. \cr
5 \tab 3 \tab \code{dispersion} \tab Estimated dispersion. \cr

6 \tab   \tab \code{null} \tab Point estimates under the null hypothesis. \cr
6 \tab 1 \tab \code{mean1} \tab Estimated mean of sample 1. \cr
6 \tab 2 \tab \code{mean2} \tab Estimated mean of sample 2. \cr
6 \tab 3 \tab \code{dispersion} \tab Estimated dispersion. \cr

7 \tab \tab \code{n1} \tab The sample size of sample 1. \cr
8 \tab \tab \code{n2} \tab The sample size of sample 2. \cr
9 \tab \tab \code{method} \tab Method used for the results. \cr
10 \tab \tab \code{ratio_null} \tab Assumed population ratio of means. \cr
11 \tab \tab \code{mle_code} \tab Integer indicating why the optimization process terminated. \cr
12 \tab \tab \code{mle_message} \tab Information from the optimizer.
}
}
\description{
Likelihood ratio test for the ratio of means from bivariate negative binomial
outcomes.
}
\details{
This function is primarily designed for speed in simulation. Missing values
are silently excluded.

Suppose \eqn{X_1 \mid G = g \sim \text{Poisson}(\mu g)} and
\eqn{X_2 \mid G = g \sim \text{Poisson}(r \mu g)} where
\eqn{G \sim \text{Gamma}(\theta, \theta^{-1})} is the random item (subject)
effect. Then \eqn{X_1, X_2 \sim \text{BNB}(\mu, r, \theta)} is the joint
distribution where \eqn{X_1} and \eqn{X_2} are dependent (though conditionally
independent), \eqn{X_1} is the count outcome for sample 1 of the items
(subjects), \eqn{X_2} is the count outcome for sample 2 of the items (subjects),
\eqn{\mu} is the conditional mean of sample 1, \eqn{r} is the ratio of the
conditional means of sample 2 with respect to sample 1, and \eqn{\theta} is
the gamma distribution shape parameter which controls the dispersion and the
correlation between sample 1 and 2.

The hypotheses for the LRT of \eqn{r} are

\deqn{
\begin{aligned}
H_{null} &: r = r_{null} \\
H_{alt} &: r \neq r_{null}
\end{aligned}
}

where \eqn{r = \frac{\bar{X}_2}{\bar{X}_1}} is the population ratio of
arithmetic means for sample 2 with respect to sample 1 and \eqn{r_{null}} is
a constant for the assumed null population ratio of means (typically
\eqn{r_{null} = 1}).

The LRT statistic is

\deqn{
\begin{aligned}
\lambda &= -2 \ln \frac{\text{sup}_{\Theta_{null}} L(r, \mu, \theta)}{\text{sup}_{\Theta} L(r, \mu, \theta)} \\
  &= -2 \left[ \ln \text{sup}_{\Theta_{null}} L(r, \mu, \theta) - \ln \text{sup}_{\Theta} L(r, \mu, \theta) \right] \\
  &= -2(l(r_{null}, \tilde{\mu}, \tilde{\theta}) - l(\hat{r}, \hat{\mu}, \hat{\theta}))
\end{aligned}
}

Under \eqn{H_{null}}, the LRT test statistic is asymptotically distributed
as \eqn{\chi^2_1}. The approximate level \eqn{\alpha} test rejects
\eqn{H_{null}} if \eqn{\lambda \geq \chi^2_1(1 - \alpha)}. However,
the asymptotic critical value is known to underestimate the exact critical
value and the nominal significance level may not be achieved for small sample
sizes. Argument \code{distribution} allows control of the distribution of
the \eqn{\chi^2_1} test statistic under the null hypothesis by use of
functions \code{\link[=asymptotic]{asymptotic()}} and \code{\link[=simulated]{simulated()}}.
}
\examples{
#----------------------------------------------------------------------------
# lrt_bnb() examples
#----------------------------------------------------------------------------
library(depower)

set.seed(1234)
sim_bnb(
  n = 40,
  mean1 = 10,
  ratio = 1.2,
  dispersion = 2
) |>
  lrt_bnb()

}
\references{
\insertRef{rettiganti_2012}{depower}

\insertRef{aban_2009}{depower}
}
\seealso{
\code{\link[=wald_test_bnb]{wald_test_bnb()}}
}
